+++ .gitignore
@@ -7,6 +7,8 @@ test/app/.sapper
 test/app/app/manifest
 test/app/export
 test/app/build
+*.js
+*.js.map
 *.ts.js
 *.ts.js.map
 !rollup.config.js
