+++ src/cli.ts
@@ -4,11 +4,7 @@ import * as child_process from 'child_process';
 import sade from 'sade';
 import * as clorox from 'clorox';
 import prettyMs from 'pretty-ms';
-import help from './cli/help.md';
-import build from './cli/build';
-import exporter from './cli/export';
-import dev from './cli/dev';
-import upgrade from './cli/upgrade';
+// import upgrade from './cli/upgrade';
 import * as ports from 'port-authority';
 import * as pkg from '../package.json';
 
@@ -18,23 +14,13 @@ prog.command('dev')
 	.describe('Start a development server')
 	.option('-p, --port', 'Specify a port')
 	.action(async (opts: { port: number }) => {
-		let port = opts.port || +process.env.PORT;
-
-		if (port) {
-			if (!await ports.check(port)) {
-				console.log(clorox.bold.red(`> Port ${port} is unavailable`));
-				return;
-			}
-		} else {
-			port = await ports.find(3000);
-		}
-
-		dev(port);
+		const { dev } = await import('./cli/dev');
+		dev(opts);
 	});
 
 prog.command('build [dest]')
 	.describe('Create a production-ready version of your app')
-	.action((dest = 'build') => {
+	.action(async (dest = 'build') => {
 		console.log(`> Building...`);
 
 		process.env.NODE_ENV = 'production';
@@ -42,52 +28,26 @@ prog.command('build [dest]')
 
 		const start = Date.now();
 
-		build()
-			.then(() => {
-				const elapsed = Date.now() - start;
-				console.error(`\n> Finished in ${prettyMs(elapsed)}. Type ${clorox.bold.cyan(dest === 'build' ? 'npx sapper start' : `npx sapper start ${dest}`)} to run the app.`);
-			})
-			.catch(err => {
-				console.error(err ? err.details || err.stack || err.message || err : 'Unknown error');
-			});
+		try {
+			const { build } = await import('./cli/build');
+			await build();
+			console.error(`\n> Finished in ${elapsed(start)}. Type ${clorox.bold.cyan(dest === 'build' ? 'npx sapper start' : `npx sapper start ${dest}`)} to run the app.`);
+		} catch (err) {
+			console.error(err ? err.details || err.stack || err.message || err : 'Unknown error');
+		}
 	});
 
 prog.command('start [dir]')
 	.describe('Start your app')
 	.option('-p, --port', 'Specify a port')
 	.action(async (dir = 'build', opts: { port: number }) => {
-		let port = opts.port || +process.env.PORT;
-
-		const resolved = path.resolve(dir);
-		const server = path.resolve(dir, 'server.js');
-
-		if (!fs.existsSync(server)) {
-			console.log(clorox.bold.red(`> ${dir}/server.js does not exist — type ${clorox.bold.cyan(dir === 'build' ? `npx sapper build` : `npx sapper build ${dir}`)} to create it`));
-			return;
-		}
-
-		if (port) {
-			if (!await ports.check(port)) {
-				console.log(clorox.bold.red(`> Port ${port} is unavailable`));
-				return;
-			}
-		} else {
-			port = await ports.find(3000);
-		}
-
-		child_process.fork(server, [], {
-			cwd: process.cwd(),
-			env: Object.assign({
-				NODE_ENV: 'production',
-				PORT: port,
-				SAPPER_DEST: dir
-			}, process.env)
-		});
+		const { start } = await import('./cli/start');
+		start(dir, opts);
 	});
 
 prog.command('export [dest]')
 	.describe('Export your app as static files (if possible)')
-	.action((dest = 'export') => {
+	.action(async (dest = 'export') => {
 		console.log(`> Building...`);
 
 		process.env.NODE_ENV = 'production';
@@ -95,21 +55,23 @@ prog.command('export [dest]')
 
 		const start = Date.now();
 
-		build()
-			.then(() => {
-				const elapsed = Date.now() - start;
-				console.error(`\n> Built in ${prettyMs(elapsed)}. Exporting...`);
-			})
-			.then(() => exporter(dest))
-			.then(() => {
-				const elapsed = Date.now() - start;
-				console.error(`\n> Finished in ${prettyMs(elapsed)}. Type ${clorox.bold.cyan(`npx serve ${dest}`)} to run the app.`);
-			})
-			.catch(err => {
-				console.error(err ? err.details || err.stack || err.message || err : 'Unknown error');
-			});
+		try {
+			const { build } = await import('./cli/build');
+			await build();
+			console.error(`\n> Built in ${elapsed(start)}. Exporting...`);
+
+			const { exporter } = await import('./cli/export');
+			await exporter(dest);
+			console.error(`\n> Finished in ${elapsed(start)}. Type ${clorox.bold.cyan(`npx serve ${dest}`)} to run the app.`);
+		} catch (err) {
+			console.error(err ? err.details || err.stack || err.message || err : 'Unknown error');
+		}
 	});
 
 // TODO upgrade
 
 prog.parse(process.argv);
+
+function elapsed(start: number) {
+	return prettyMs(Date.now() - start);
+}
